+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Image processing now supports WebP
 - Fix `zola serve` failing for some static files
 - Fix `zola serve` not picking up directory renaming
+- Add `path` to the taxonomy terms to be on par with pages and sections
 
 ## 0.13.0 (2021-01-09)